Added Precompiled Header support && Changed keyword
Added Precompiled Header support - I know cmake allows multiple PCHs,
however this isn't a universal thing, I want this application to cater
to as many build systems in the future, so this is permanent.

Changed Keyword - Changed ROOTDIR to WORKSPACEDIR as it relates to where
the workspace in prebuild is going to be located. Naming will be like
this going forward